ENH: Add CoW optimization to interpolate
